full worked int

+++ includes/ft_printf.h
@@ -48,6 +48,6 @@ typedef	struct	s_print
 int				ft_printf(const char *format, ...);
 void			ft_checkarg(va_list ap, char *str);
 int				ft_print_s(char *str, char *new);
-int				ft_print_di(int nbr, char *new);
+int				ft_print_di(long long int nbr, char *new);
 
 #endif
